Address remaining blocks.
This commit takes a stab at addressing the remaining block margins. Remember, the goal is for the editor to be as easy to style as possible, so that it's simpler to make it look like the frontend. Before this PR, margins were all the same, and applied in a sweeping fashion. This PR puts the onus on each and every block and HTML element, to define their own margins, and only when it's actually necessary to do so.

Why is it ever necessary? Because otherwise, CSS bleeds from wp-admin into the iframe. As a reminder, we will continue to work to refactor that to a state where CSS does not bleed into the editing canvas. But in the mean time, this is a step to improving things.

In that vein, it addresses all blocks now present in Gutenberg, including placeholders, readmore, nextpage, and other meta-blocks.

- Some have their margins stored in style.scss because it is structural to the block, like button.
- Some have their margins stored in editor.scss because it is only necessary in the editor, like the placeholder.
- Some have their margins stored in theme.scss, because there is no CSS bleed, and that way we can consider it an "opinionated style", like the Code block.
- Others have their style stored in the new normalization stylesheet, editor-styles.scss, which is the one we hope to eventually not need at all.

Please share your thoughts and let's discuss.